pkg/sysinfo/checkCgroupPids: speedup
For some reason this code chose not to use information already fetched,
and call cgroups.FindCgroupMountpoint() instead. This is not a cheap
call, as it has to parse the whole nine yards of /proc/self/mountinfo,
and the info it tries to get (whether the pids controller is present)
is already available from cgMounts map.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>