Add script to rebuild images on quay.io
We recently discovered that the build triggers that are employed on
quay.io for the Buildah, Podman, and Skopeo container images does not
refresh the images as frequently as we thought.  Due to Docker using
the `--no-cache` flag under the covers at quay to do the builds, our
images were always using the pre-existing cache unless the `FROM` target
or the Dockerfile changed.  This rarely happened and it would take
weeks or more before the stable version would bump up to the latest
version available.  Ditto the upstream and testing variants.

I've created a script that can be run `by-hand` or via cron to update the
various images.  The script takes the project name and the creds for
an administrator on our repos on quay.io.  If those values are not supplied
the script will ask for them and the password will not be shown when
typed in.

I'm planning to run this via cron on one of my VM's and the images should
be updated every morning that my VM is up, or someone else with admin privs
can run it.  If someone runs this script without privs, they will build the
images locally and they will not be able to push them.

This was the most logical place I could think of to drop this script as it's
usable by all of our projects.  If you've a better location, please let me
know.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>