rm VERSION: version info is in the version package
Note that this change is required to be able to vendor the code.
For unknown reasons, go will reject taking the code from the
current branch while it happily eats the one from master when
vendoring into another prohject.

The full error message is:

 > server response: not found: create zip: case-insensitive file name collision: "VERSION" and "version"

Removing the redundant VERSION file resolves the issue.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>