capabilities: add new method BoundingSet()
add a new public function to retrieve all the capabilities in the
current bounding set.

This is useful for Podman to use only these capabilities when running
with --privileged as it can break running in a container where the
available capabilities can be a subset of the ones available in the
kernel.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>