Add a default network creation package
This package exposes a function for creating a default CNI
network configuration that can be used by Podman and Buildah.
This will allow us to stop shipping a default configuration file,
and instead create it on demand (once per Podman installation.
This allows for greater configurability by the user - set the
default subnet (and, later, default IPv6 subnet) in
containers.conf, and the config file will be created with those
settings.

Signed-off-by: Matthew Heon <mheon@redhat.com>