Mount the whole crypto-policies directory
To correctly determine the crypto-policy being used
in the container with the update-crypto-policies command
it needs access to durectories such as python under
/usr/share/crypto-policies. Since we were only bind mounting
the FIPS/backends directory, it was returning that the DEFAULT
policy is being used in the container even though the container
is actually in FIPS mode.
Change the bind mount to mount /usr/share/crypto-policies to
/etc/crypto-policies in the container and the correct crypto-policy
is returned with the update-crypto-policies --show command.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>