libimage: block on events
Debugging containers/podman/pull/10219 has revealed a number of issues
with respect to the consistency of events.  At the core, some events may
not be written to the backend when the runtime is shutdown.

While there are various changes required to tackle the consistency
issues, a first step is to make sure that writes to the events channel
block.

Set the timeout to 10 seconds.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>