Ensure configuration directory is created for networks
We need to make sure the default network directory gets created
if it does not exist. We previously did this after CNI was
initialized but this could race on CNI recognizing the new
network, so we moved it before CNI init. Unfortunately, CNI init
made the configuration directory, so we now have to do that
ourselves.

Signed-off-by: Matthew Heon <mheon@redhat.com>