libimage: PullCandidates
Add a new `PullCandidates()` API to query the exact candidates that
would be pulled (in order) during an image pull.

This functionality is needed in podman-remote for prompting.  Doing the
resolution in libimage keeps the logic at one place and makes sure that
the exact same code is used as it would during a pull.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>