Suppress staticcheck linter warnings
Fix a couple of warnings like this one:

> pkg/formats/templates.go:23:14: SA1019: strings.Title is deprecated:
> The rule Title uses for word boundaries does not handle Unicode
> punctuation  properly. Use golang.org/x/text/cases instead.
> (staticcheck)

This is caused by the fact that Go 1.18 strings package made a
(somewhat strange) decision to deprecate strings.Title. The suggestion
to replace it with a new package is a bit too much.

Silence the warning instead, as we are pretty sure this function works
for us.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>