Define default loggers in man page
We have received two Bugzilla's about the upstream
documentation being incorrect about the default value of the
events_logger field in containers.conf.  The are:

https://bugzilla.redhat.com/show_bug.cgi?id=2076664
https://bugzilla.redhat.com/show_bug.cgi?id=2076665

From what I can see, we use `file` in RHEL and the podman machine,
in all other distributions were' using journald.  I've
attempted to change the man page to reflect that.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>