CI: enable extra linters for new code
The idea here is to run some more picky linters only on new code,
so we can keep the existing code as is, but impose some additional
requirements to any new code.

Currently, this enables godot linter which wants all comments to have a
period at the end of the sentence. This makes sense because otherwise
some comments might be read as unfinished.

Obviously, no one wants to fix all the existing comments to have
periods, but this will impose such requirement onto any new code
(which also includes the code that is moved around).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>