libnetwork: add a dummy network backend
Add a dummy network backend that fails on every operation requested to
it.

This allows consumers of this package to obtain and carry around (as
long it's not really used) a types.ContainerNetwork object even on
targets where no network backend is supported, which is convenient to
avoid having to gate a significant amount of code behind run-time
and/or build-time conditionals.

Signed-off-by: Sergio Lopez <slp@redhat.com>