Add option to push all tags
This seeks to mirror `docker push --all-tags IMAGE`.

Because tags are appended to the destination, this will only
work with the docker transport. Otherwise you'd get directories
overwriting or with weird names.

Also note that if AllTags is true then the user must provide the
name of the image only; providing a tag will crash.

Docker has this behavior:
```
tag can't be used with --all-tags/-a
```

Requirement for https://github.com/containers/podman/pull/14949

Signed-off-by: Brian Yarbrough <bcynmelk+git@gmail.com>