Improve push AllTags logic
1. Inverts test of AllTags bool, so that false is handled first the
single image is pushed.
2. Makes use of library functions to test source formats instead of
flaky string parsing.
3. Ensures that all tags across mulitple images for a given repo are
pushed, instead of just for the single image.
4. Make the AllTags bool and a specified destination mutually exclusive.

Note that Podman currently assumes the destination is the source
if no destination is provided, rather than leaving it as an empty string.
This will cause an error, and I think should be fixed in a PR to that repo
since libimage.PushImage already fills in resolvedSource for an
empty destination.
https://github.com/containers/podman/blob/v4.2/cmd/podman/images/push.go#L131

Also note that the unit test is not able to push images to a real docker
registry, so I would consider them incomplete, but don't know how to improve.

Signed-off-by: Brian Yarbrough <bcynmelk+git@gmail.com>