pkg/randid: add
As noted by comments in pkg/configmaps and pkg/secrets, the string
returned from stringid package is 64 bytes long, and had to be cut down
to 25 bytes. This essentially means that we do some extra work for
nothing.

Also, since we always check that the string is unique, we don't need any
fancy randomness and can do with math.Rand easily.

Replace using stringid with a small in-house implementation.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>