fix netavark IgnoreIfExists config test flake
We cannot compare network with Expect().To(Equal()), the internal
time.Time is not equal after parsing from file. There is some internal
field offset that can be different but the date is still the same. Other
tests have the same problem and to circumvent this we use EqualNetwork()
there. Do the same here.

Fixes #1335

Signed-off-by: Paul Holzinger <pholzing@redhat.com>