libnetwork/netavark: pass plugin directories to nv
Netavark needs to get the actual list of directories from podman because
users can set them in contianers.conf.

Note that this change requires a netavark with
https://github.com/containers/netavark/pull/509
To allow better backwards compat we only pass the new argument when we
need. Also added the missing ipvlan driver to the supported driver list.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>