Relax name length restriction for Secrets/ConfigMaps.
The Kubernetes documentation states that the name of these resources
much be a DNS subdomain name, which it defines[1] as being up to 253
characters.

Following the referenced RFC it does appear that the name should be
further restricted to require the 253 characters be made up of multiple
of multiple 63-character labels separated by dots. The Kubernetes[2]
validation code does not apply this label restriction though, and will
accept names longer than 63 characters.

This commit adjusts the length limits to match.

[1]: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#dns-subdomain-names
[2]: https://github.com/kubernetes/apimachinery/blob/6b1428efc73348cc1c33935f3a39ab0f2f01d23d/pkg/util/validation/validation.go#L215

Signed-off-by: Kenny MacDermid <kenny@macdermid.ca>