libnetwork: add pasta package to start pasta
Mostly copied from podman/libpod/networking_pasta_linux.go, this should
be moved to common so that we can use it in buildah as well.

I changed the function parameter to accept a single struct with all
options so that we can extend it when needed without breaking the API
The Container type is libpod/podman specific and not available here so
we only use the values we really need here.
Also I removed "pasta" from the function and constant name as this will
now be used as pasta.Setup() so it would be redundant.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>