libnetwork/pasta: fix possible panic in error cast
exec.Command().Output() can fail for several errors, i.e. maybe it was
not able to execute at all. We only get a error from type
`exec.ExitError` when the command exited > 0. For all other cases this
cast was unsafe and would cause a panic with a nil deref.

First check the error type correctly with errors.As() then return the
stderr and exit code. Otherwise just return the wrapped error as is.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>