add pkg/rootlessport from podman
This will be used by the slirp4netns code I plan to port to c/common.
The actual rootlessport binary will stay in podman this is just so we
can send the config to stdin from here.

The code was copied from podman and I only altered the top level comment
as this no longer applies.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>