add slirp4netns package
Import code from podman/libpod/networking_slirp4netns.go.

This code should live here so we can share it with buildah later.
The code here was heaviliy modified in order to compile because we
cannot use the Container struct here. This should make the code better
separated from the libpod internals and allows it to be used by other
callers, i.e. buildah.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>