mac helper binaries: add default brew libexec path
After installing podman via brew, I expected a locally compiled one to
work as well.  But it failed as it could not find gvproxy.  It turns out
that brew installs it at /opt/homebrew/opt/podman/libexec/podman/gvproxy,
so I want the directory to be part of the default look-up paths.

Same applies to the `/usr/local/opt` paths used on Intel Macs.  Also
rearrange the order a bit.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>