pkg/subscriptions: add /etc/crypto-policies/config in fips mode
Make sure /etc/crypto-policies/config is bind mounted from the host so
it does contain the proper FIPS value and a reinstall of
crypto-policies-scripts does not overwrite the existing files with the
default config.

We create a tmpfile and write FIPS to it so we can bind mount it and the
contianer can change the config if needed.

Fixes https://issues.redhat.com/browse/RHEL-9836

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit 417611c5d18f44470e1b15fc8dc36adea7692edf)