libnetwork/cni: ignore ENOENT while reading networks
At this point we already read the config file but here we have to get
timestamp here so we can hit another ENOENT if it was removed in the
meantime. Just ignore this and do not log an error as this is normal
behavior when another process is deleting a network in parallel.

Fixes containers/podman#20173

Signed-off-by: Paul Holzinger <pholzing@redhat.com>