Document GetDefaultAuthFile, and don't call it in CLI defaults
We _need_ to know whether the user explicitly set an option
or not, because --authfile and --compat-auth-file conflict.

Also, --authfile= (unset) and --authfile=$the_default_auth_json
have different effects, so setting the default value of the option
is _not_ a practical way to show what the system would do if the
option were not specified.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>