pkg/manifests.list.AddInstance(): don't set Platform to an empty struct
If the Platform struct for an OCI instance being added to an image index
is empty, set the Platform pointer field in the Descriptor struct to
`nil` instead of setting it to point to a Platform struct filled with
empty values.  I expect this to be the case for at least some OCI
artifacts being added to indexes.

Likewise, if we're updating the Platform for an OCI instance and its
value goes to zero, set its pointer to `nil` to avoid the same case.

When reading information back from a descriptor, or setting a field in
the Platform struct, don't assume that Platform is (already) set to a
non-nil value.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>