pkg/manifests.list.preferOCI() learn about some new OCI fields
The OCI image index structure has grown a few new fields since the last
time we looked, so we should be paying attention to whether any of them
are being used when deciding if we need to use the OCI format instead of
the Docker format.

The Docker format also grew a "URLs" field, so we can't use its mere
presence as a heuristic any more.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>