Handle multiple artifacts correctly
When saving artifacts to OCI layout directories to use as supplemental
refernces, use a different directory for each artifact(!)

When supplementing a reference in CopySpecificImages mode, only mark a
supplemental reference as the source for a specified instance if it
actually corresponds to the instance's digest.

If an artifact manifest doesn't have any file layers, remember to
catalog the empty descriptor that we add to its layer list to keep it
from being empty, as recommended.

Extend libimage/manifests.TestReference() to ensure that we're able to
copy everything that we should be able to copy.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>