pkg/config: remove deprecated MachineEnabled field
We now only use the marker file, see pkg/machine.
The advantage of this is that pkg/machine no longer imports pkg/config
and we use it in rootlessport which means it will debloat the binary
there significantly.

The file approach has been added in podman 4.1 (2 years ago) so
basically all machines should have it now and this shouldn't break
anything.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>