pkg/config: make EditConnectionConfig race free
Add a lock to EditConnectionConfig to make it race free for parallel
modification. It is possible that several podman system connection
commands are called simultaneously so we should read/file in a
consistent way to ensure no modifications are "silently dropped".

A test is added that reliably fails without the lock.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>