libnetwork/pasta: add new Setup2 to return result
Currently both callers in podman and buildah join and inspect the netns
to get the local ip configured by pasta in order to add it to
/etc/hosts. So instead of doing this in two places let's just do it here
once and return the result to the caller.

In order to not cause vendoring issues I decided against breaking the
API and added a new Setup2 function instead. I will then update
podman and buildah to make use of it.

Also I plan on adding more fields in the result, i.e. dns address.

Because this now depends on linux only functionality make sure to only
build it on linux, pasta only works on linux anyway so this is not a
problem.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>