libnetwork/pasta: set --dns-forward as default
By default set 169.254.0.1 as nameserver in the container, right now we
do not do special dns handling which means if a user has only localhost
resolver or the same nameserver ip as the host ip used by pasta then dns
will most likely fail.

pasta allows us to remap one ipv4 for dns which will then automatically
get remapped to the host dns server from resolv.conf. For that we must
use --dns-forward, now the choice of which ip is arbitrary but using the
local link address 169.254.0.1 is unlikely to be used so it should avoid
conflicts.

Also return the ip in the result together with a ipv6 bool so that
podman can create a correct resolv.conf with that ip for the container.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>