libnetwork/util: add GetLocalIPExcluding()
For some cases it may be required to get a local ip while excluding a
certain ip. This will be used to set a better host.containers.internal
name for pasta networks as it is likely that host eth0 ip == container
ip so in this case we must use different host ip.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>