revert dep bumps that require go 1.21
So apparently go 1.21 introduced a new toolchain directive that
specifies the minimum version that go uses to build. Also the go
directive is now a hard requirement. We are not ready for this at all,
podman CI still uses f38 with go 1.20.

We need to figure a proper way how to deal with this, for now do not
merge any PR's that set toolchain or bump the go version to 1.21.

see https://go.dev/doc/toolchain

Signed-off-by: Paul Holzinger <pholzing@redhat.com>