Add a libimage.Runtime.LoadReference for loading images
Add a LoadReference() method, for when we already know what format the
local image should be treated as, and we don't have the patience for
Load()'s guessing at it.

Teach libimage.Runtime.copyFromDefault() about the docker transport.
It's already required at build-time because it's directly referenced
elsewhere in the same file.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>