libnetwork/netavark: fix rootless unit tests
Commit 26fc823c27 added check for the env var inside the network
interface setup code as this is something that is always set by podman.
However if you try to run the unit tests as rootless they now always
failed as the code assumed we have access to a rootful path.

The easy fix is to just fix the test to set the env when running
rootless.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>