[skip-ci] RPM: ghost `/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-release`
Packages built on CentOS Stream 10 aren't installable on RHEL 10 because
of installation conflict on /etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-release.
So far, that has prevented us from early testing of packages on proper
RHEL 10 environments.

This will continue to be an issue when an upcoming RHEL is available for
testing but doesn't have copr build targets enabled and you need to
depend on the equivalent CentOS Stream targets.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>