libimage: recommend podman system check over reset
podman system reset nukes everything which makes not a great suggestion.
Now that we have the podman system check recommend that instead as it
actually checks the storage for corruption and might be able to repair
it.

This is important because as it turns out this warning can pop up even
in totally normal circumstances as there is a race between image and
layer listing[1]. And even if that get fixed the check command will be
the better suggestion.

[1] https://github.com/containers/podman/issues/23331

Signed-off-by: Paul Holzinger <pholzing@redhat.com>