pasta: update warnings line
Use %s as %q just quotes/escapes everything which makes it harder to
read and trim of the last newline and spaces as well.

Also update the warnings comment, we still see warnings by default on
our debian VMs in podman CI so this cannot be on the warning level yet.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>