[v0.51] exclude github.com/mitchellh/osext
Exclude the old, outdated, and problematic
github.com/mitchellh/osext which was being dragged in
by containerd.

This is what containerd did in https://github.com/containerd/containerd/pull/10011
and included it in 1.6.31.  I tried bumping first to 1.6.31, but
that dragged in a lot of stuff.  I think this is safer, if it works.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>