update CONTRIBUTING.md
Change the DCO wording so that it is clear that we do not need the real
name, just someting the properly identifies that person within the
community.

Touch up the old Reporting Issues and Submitting Pull Requests section
based on Matt's work in podman[1] and add some extra step in there.

Add new sections for the vendor process and and how to use git bisect.

[1] https://github.com/containers/podman/pull/24261

Signed-off-by: Paul Holzinger <pholzing@redhat.com>